different bank account works, made max borrowable virtual. fixed a bug in current account cpp,
allow the user to select different types of bank account (except saving and standard bank account).

made max borrowable virtual so that show all available funds works correctly. NOTE: THE CHILD ACCOUNT, I'M NOT SURE ON HOW TO CALCULATE THIS.

fixed a bug in the current account which was calling the wrong function (getdata from stream) instead of calling the correct function (getAccountDataFromStream).

NOTE: CAN WIDTHDRAW AND MAX BORROWABLE MAY NEED TO BE IMPLEMENTED IN THE CHILD/CURRENT CLASS.